move endpoints to own file, rename tests, add tests for client
