+++ chessdotcom/aio.py
@@ -1,4 +1,4 @@
-from chessdotcom import Client
-from chessdotcom.client import *
+from .client import Client
+from .endpoints import *
 
 Client.aio = True
