+++ pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
-testpaths =
-    tests
-    
+testpaths = tests
+addopts = -p no:warnings
+
